+++ packages/user-management/spec/ut/standard-user-workflow/legacy/LegacyAdapterUserRepository.spec.ts
@@ -1,11 +1,6 @@
 import 'jasmine';
 
-import {
-  LegacyAdapterUserRepository,
-  StandardAccountStatus,
-  LoginAlreadyUsedError,
-  LegacySimpleError
-} from '../../../../src';
+import { LegacyAdapterUserRepository, StandardAccountStatus, LegacyLoginAlreadyUsedError } from '../../../../src';
 import { IllegalStateError } from '../../../../src/common/api';
 
 describe(`LegacyAdapterUserRepository`, () => {
@@ -88,7 +83,7 @@ describe(`LegacyAdapterUserRepository`, () => {
         // THEN
         expect(() => {
           throw e;
-        }).toThrowError(LoginAlreadyUsedError, 'Login "odile.deray" is already used');
+        }).toThrowError(LegacyLoginAlreadyUsedError, 'Login "odile.deray" is already used');
       }
     });