+++ packages/user-management/src/standard-user-workflow/configurer/index.ts
@@ -2,3 +2,4 @@ export * from './account';
 export * from './confirmation';
 export * from './DefaultStandardWorkflowConfigurer';
 export * from './StandardUserWorkflowConfigurer';
+export * from './lost-password';