+++ packages/user-management/src/standard-user-workflow/core/index.ts
@@ -3,3 +3,4 @@ export * from './authentication';
 export * from './exceptions';
 export * from './StandardUserWorkflow';
 export * from './StandardUserWorkflowAsyncTaskService';
+export * from './lost-password';