+++ packages/user-management/src/standard-user-workflow/core/lost-password/index.ts
@@ -1,0 +1,2 @@
+export * from './EmailAskResetPasswordManager';
+export * from './ConfirmResetPasswordManager';